Manage

+++ src/com/kh/theater/model/dao/TheaterDao.java
@@ -43,8 +43,8 @@ public class TheaterDao {
 			while(rset.next()) {
 				list.add(new Theater(rset.getInt("THEATER_NO"), rset.getString("NAME"),
 						rset.getString("ADDRESS"), rset.getString("PHONE"), rset.getInt("ROOM_COUNT"),
-						rset.getString("TRANSPORT"), rset.getString("PARKING"), rset.getInt("LONGITUDE"),
-						rset.getInt("LATITUDE"), rset.getInt("SECTION_NO")));
+						rset.getString("TRANSPORT"), rset.getString("PARKING"), rset.getDouble("LONGITUDE"),
+						rset.getDouble("LATITUDE"), rset.getInt("SECTION_NO")));
 			}
 		} catch (SQLException e) {
 			e.printStackTrace();
@@ -70,8 +70,8 @@ public class TheaterDao {
 			while(rset.next()) {
 				list.add(new Theater(rset.getInt("THEATER_NO"), rset.getString("NAME"),
 						rset.getString("ADDRESS"), rset.getString("PHONE"), rset.getInt("ROOM_COUNT"),
-						rset.getString("TRANSPORT"), rset.getString("PARKING"), rset.getInt("LONGITUDE"),
-						rset.getInt("LATITUDE"), rset.getInt("SECTION_NO")));
+						rset.getString("TRANSPORT"), rset.getString("PARKING"), rset.getDouble("LONGITUDE"),
+						rset.getDouble("LATITUDE"), rset.getInt("SECTION_NO")));
 			}
 		} catch (SQLException e) {
 	
@@ -98,7 +98,7 @@ public class TheaterDao {
 			if(rset.next()) {
 				t = new Theater(rset.getInt("THEATER_NO"), rset.getString("NAME"), rset.getString("ADDRESS"),
 						rset.getString("PHONE"), rset.getInt("ROOM_COUNT"), rset.getString("TRANSPORT"),
-						rset.getString("PARKING"), rset.getInt("LONGITUDE"), rset.getInt("LATITUDE"),
+						rset.getString("PARKING"), rset.getDouble("LONGITUDE"), rset.getDouble("LATITUDE"),
 						rset.getInt("SECTION_NO"));
 			}
